care about count of spams

+++ core/src/main/java/mastodon4j/api/entity/PollOption.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 class PollOption(
     @SerializedName("title") val title: String = "",
-    @SerializedName("votes_count") val votesCount: Int? = 0
+    @SerializedName("votes_count") val votesCount: Long? = 0
 )
